Add files via upload
The add and search functions works, but I am stuck with the removing function. I am attempting to delete a node and replace it with a successor through pointers. In my find_node() function, I'm not sure why the "save_prev" node updates when the "current" node address changes. This defeats the purpose of tracking the previous node's memory address.